+++ packages/sam/src/builders/cloudformation/deploy/deploy.ts
@@ -1,14 +1,10 @@
-import { createBuilder, BuilderContext } from '@angular-devkit/architect';
+import { createBuilder } from '@angular-devkit/architect';
 import { JsonObject } from '@angular-devkit/core';
-import { underscore } from '@angular-devkit/core/src/utils/strings';
 import { runCloudformationCommand } from '../run-cloudformation-command';
-import { loadCloudFormationTemplate } from '../../../utils/load-cloud-formation-template';
 import { CloudFormationDeployOptions } from './CloudFormationDeployOptions';
-import {
-    ImportStackOutputs,
-    formatStackName,
-    OutputValueRetriever,
-} from '@nx-aws/core';
+import { ImportStackOutputs, formatStackName } from '@nx-aws/core';
+import { IParameterOverrides } from './IParameterOverrides';
+import { getParameterOverrides } from '../../../utils/getParameterOverrides';
 
 export type Capability =
     | 'CAPABILITY_IAM'
@@ -58,10 +54,6 @@ interface IDeployOptions extends JsonObject {
     stackSuffix: string | null;
 }
 
-export interface IParameterOverrides {
-    [key: string]: string;
-}
-
 try {
     require('dotenv').config();
 } catch (e) {}
@@ -99,54 +91,3 @@ export default createBuilder<IDeployOptions>(async (options, context) => {
     };
     return runCloudformationCommand(cfOptions, context, 'deploy');
 });
-
-async function getParameterOverrides(
-    options: IDeployOptions,
-    context: BuilderContext,
-    stackSuffix: string | undefined
-): Promise<IParameterOverrides> {
-    const cf = loadCloudFormationTemplate(options.templateFile);
-    const parameters = cf.Parameters;
-    if (!parameters) {
-        return {};
-    }
-    const importStackOutputs = options.importStackOutputs;
-    const overrides: IParameterOverrides = {};
-    if (importStackOutputs) {
-        const outputValueRetriever = new OutputValueRetriever();
-        // retrieve the values from the other projects
-        const values = await outputValueRetriever.getOutputValues(
-            importStackOutputs,
-            context,
-            stackSuffix
-        );
-        options.parameterOverrides = {
-            ...(options.parameterOverrides || {}),
-            ...values,
-        };
-    }
-    for (const key in parameters) {
-        if (parameters.hasOwnProperty(key)) {
-            if (
-                options.parameterOverrides &&
-                options.parameterOverrides.hasOwnProperty(key)
-            ) {
-                overrides[key] = options.parameterOverrides[key];
-            } else {
-                const envKey = underscore(key).toUpperCase();
-                const value = process.env[envKey];
-                if (value) {
-                    context.logger.info(
-                        `Retrieved parameter override ${key} from environment variable ${envKey}`
-                    );
-                    overrides[key] = value;
-                } else if (!parameters[key].Default) {
-                    context.logger.fatal(
-                        `Missing parameter override ${key}; deploy will likely fail`
-                    );
-                }
-            }
-        }
-    }
-    return overrides;
-}