+++ packages/sam/src/builders/cloudformation/package/mapRelativePathsToAbsolute.ts
@@ -11,42 +11,46 @@ const MAPPINGS: Record<string, ResourcePropertiesToMap | undefined> = {
     'AWS::ApiGateway::RestApi': { properties: ['BodyS3Location'], skip: false },
     'AWS::Lambda::Function': { properties: ['Code'], skip: true },
     'AWS::Serverless::Function': { properties: ['CodeUri'], skip: true },
+    'AWS::Serverless::LayerVersion': {
+        properties: ['ContentUri'],
+        skip: false,
+    },
     'AWS::AppSync::GraphQLSchema': {
         properties: ['DefinitionS3Location'],
-        skip: false
+        skip: false,
     },
     'AWS::AppSync::Resolver': {
         properties: [
             'RequestMappingTemplateS3Location',
-            'ResponseMappingTemplateS3Location'
+            'ResponseMappingTemplateS3Location',
         ],
-        skip: false
+        skip: false,
     },
     'AWS::Serverless::Api': { properties: ['DefinitionUri'], skip: false },
     'AWS::Include': { properties: ['Location'], skip: false },
     'AWS::ElasticBeanstalk::ApplicationVersion': {
         properties: ['SourceBundle'],
-        skip: false
+        skip: false,
     },
     'AWS::CloudFormation::Stack': { properties: ['TemplateURL'], skip: false },
     'AWS::Serverless::Application': { properties: ['Location'], skip: false },
     'AWS::Serverless::StateMachine': {
         properties: ['DefinitionUri'],
-        skip: false
+        skip: false,
     },
-    'AWS::Glue::Job': { properties: ['Command.ScriptLocation'], skip: false }
+    'AWS::Glue::Job': { properties: ['Command.ScriptLocation'], skip: false },
 };
 
 export function mapRelativePathsToAbsolute(
     template: Template,
     inputDir: string
-) {
+): void {
     const resources = template.Resources;
     if (!resources) {
         return;
     }
     for (const name in resources) {
-        if (resources.hasOwnProperty(name)) {
+        if (Object.prototype.hasOwnProperty.call(resources, name)) {
             const resource = resources[name];
             const properties = resource.Properties;
             const mapping = MAPPINGS[resource.Type];