+++ CHANGELOG.md
@@ -2,6 +2,39 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.11.1](https://github.com/studds/nx-aws/compare/v0.11.0...v0.11.1) (2021-03-03)
+
+
+### Bug Fixes
+
+* **sam:** correct required properties in build schema ([8aa684a](https://github.com/studds/nx-aws/commit/8aa684a5e154d5eb5198bfa79f8c90e165845e52))
+
+
+
+# [0.11.0](https://github.com/studds/nx-aws/compare/v0.10.0...v0.11.0) (2021-03-01)
+
+
+### Features
+
+* **sam/build:** install npm modules if generatePackageJson is on ([a93e230](https://github.com/studds/nx-aws/commit/a93e23066e7c1fae58ad840565cf727b58ee8647))
+* copy & paste the latest @nrwl/node:build ([773e5ce](https://github.com/studds/nx-aws/commit/773e5ce1085c25d64b6fb62b8ad2a40dc40a59a9))
+
+
+
+# [0.10.0](https://github.com/studds/nx-aws/compare/v0.9.1...v0.10.0) (2021-02-24)
+
+
+### Bug Fixes
+
+* get out of the way of how sam-cli handles layers [#55](https://github.com/studds/nx-aws/issues/55) ([6632316](https://github.com/studds/nx-aws/commit/6632316ad0283b5aeffa80912b083e0d3b66ef24))
+
+
+### Features
+
+* **sam:** add support for parameter overrides to execute ([7d9bfbf](https://github.com/studds/nx-aws/commit/7d9bfbf7441b48b26441589e7038e25fb71c7274))
+
+
+
 ## [0.9.1](https://github.com/studds/nx-aws/compare/v0.9.0...v0.9.1) (2021-02-24)