+++ e2e/s3-e2e/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.11.1](https://github.com/studds/nx-aws/compare/v0.11.0...v0.11.1) (2021-03-03)
+
+
+
+# [0.11.0](https://github.com/studds/nx-aws/compare/v0.10.0...v0.11.0) (2021-03-01)
+
+
+
+# [0.10.0](https://github.com/studds/nx-aws/compare/v0.9.1...v0.10.0) (2021-02-24)
+
+
+
 ## [0.9.1](https://github.com/studds/nx-aws/compare/v0.9.0...v0.9.1) (2021-02-24)